% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/act_tbl.R
\name{act_tbl-class}
\alias{act_tbl-class}
\title{\code{act_tbl} class}
\description{
The \code{act_tbl} S3 class is a subclass of \code{\link[base:data.frame]{data.frame}}
and \code{\link[tibble:tbl_df-class]{tibble}}.
}
\details{
In nearly every respect, it can be treated like a tibble; however, this
allows the package to provide an improved \code{summary.act_tbl()} function
to get an overview of the activity.
}
