% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activatr.R
\docType{package}
\name{activatr}
\alias{activatr}
\title{activatr: Utilities for Parsing and Plotting Activities}
\description{
This contains helpful functions for parsing, managing, plotting, and
visualizing activities, most often from GPX (GPS Exchange Format) files
recorded by GPS devices. It allows easy parsing of the source files into
standard R data formats, along with functions to compute derived data for
the activity, and to plot the activity in a variety of ways.
}
