% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{get_ggmap_from_df}
\alias{get_ggmap_from_df}
\title{Get a ggmap object for a given Activatr DF.}
\usage{
get_ggmap_from_df(df, ...)
}
\arguments{
\item{df}{A Activatr DF: a tibble from \code{parse_gpx} or \code{parse_tcx}.}

\item{...}{Additional arguments to pass to \code{ggmap::get_googlemap}.}
}
\value{
A ggmap object, the result of calling \code{ggmap::get_googlemap},
  but with the correct center and size to include the entire data frame.
}
\description{
Note that since this calls \code{ggmap::get_googlemap}, you must have
previously called \code{ggmap::register_google} to register an API key.
}
