% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.R
\name{pace_formatter}
\alias{pace_formatter}
\title{A formatter that takes a pace duration and returns a formatted M:SS string.}
\usage{
pace_formatter(pace)
}
\arguments{
\item{pace}{a lubridate duration.}
}
\value{
a formatted string representing the pace.
}
\description{
A formatter that takes a pace duration and returns a formatted M:SS string.
}
\examples{
pace_formatter(lubridate::dseconds(390))
}
