% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{running_example_ggmap}
\alias{running_example_ggmap}
\title{The result of calling get_ggmap_from_df on running_example}
\format{
An object of class \code{ggmap} (inherits from \code{raster}) with 1280 rows and 1280 columns.
}
\usage{
running_example_ggmap
}
\description{
This is the result of running:
}
\details{
```
running_file <- system.file(
  "extdata",
  "running_example.gpx",
  package = "activatr")
running_df <- parse_gpx(running_file)
running_example_ggmap <- get_ggmap_from_df(running_df)
```

except using that in vignettes or examples is hard, because
\code{get_ggmap_from_df} requires an api key be passed to `ggmap`.
So this is the result of running that with a valid API key.
}
\keyword{datasets}
