% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{running_example_ggmap}
\alias{running_example_ggmap}
\title{Precomputed example ggmap}
\format{
An object of class \code{ggmap} (inherits from \code{raster}) with 1280 rows and 1280 columns.
}
\usage{
running_example_ggmap
}
\description{
This is a precomputed ggmap for running_example.gpx, to avoid needing an
API key in vignettes, which \code{get_ggmap_from_df()} requires.
}
\details{
This is the result of running:

\if{html}{\out{<div class="sourceCode">}}\preformatted{running_file <- system.file(
  "extdata",
  "running_example.gpx",
  package = "activatr")
running_df <- parse_gpx(running_file)
running_example_ggmap <- get_ggmap_from_df(running_df)
}\if{html}{\out{</div>}}

after setting a valid API key.
}
\keyword{internal}
