% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.R
\name{speed_to_mile_pace}
\alias{speed_to_mile_pace}
\title{Converts a speed (in meters per second) to a mile pace}
\usage{
speed_to_mile_pace(speed)
}
\arguments{
\item{speed}{a vector of speed values in meters per second,
as from \code{mutate_with_speed}.}
}
\value{
a corresponding vector of lubridate durations, representing the
  mile pace.
}
\description{
Converts a speed (in meters per second) to a mile pace
}
\examples{
speed_to_mile_pace(1)
}
